Move firebase_service.dart
